day_14_langchain_basics/requirements.txt
langchain
langchain-mistralai
langchain-community
langchain-openai
langchain-ollama
langchainhub
python-dotenv
langchain-mistralai